server/requirements.txt
pymongo==4.3.3
selenium==4.9.0
python-dotenv==1.0.0